requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cloudinary==1.44.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
